feat: created config files and moved every constant into the corresponding config_file
