+++ nimsort_logic/configs/config_axis.py
@@ -1,0 +1,29 @@
+# axis_controller_node
+MAX_VELOCITY_X = 0.05
+MAX_VELOCITY_Y = 0.05
+MAX_VELOCITY_Z = 0.05
+MAX_ACCELERATION_X = 0.02
+MAX_ACCELERATION_Y = 0.02
+MAX_ACCELERATION_Z = 0.02
+POSITION_TOLERANCE_X = 0.001
+POSITION_TOLERANCE_Y = 0.001
+POSITION_TOLERANCE_Z = 0.001
+VELOCITY_TOLERANCE_X = 0.001
+VELOCITY_TOLERANCE_Y = 0.001
+VELOCITY_TOLERANCE_Z = 0.001
+KP_X = 1.2
+KP_Y = 1.2
+KP_Z = 1.2
+KD_X = 2.3
+KD_Y = 2.3
+KD_Z = 2.3
+D_FILTER_ALPHA_X = 0.5
+D_FILTER_ALPHA_Y = 0.5
+D_FILTER_ALPHA_Z = 0.5
+TF_X = 0.00875
+TF_Y = 0.00875
+TF_Z = 0.00875
+
+# init_process
+ZERO_ACCELERATION = (0.0, 0.0, 0.0)
+HOMING_ACCELERATION = (0.005, 0.005, -0.01)
