+++ nimsort_logic/configs/config_position_prediction.py
@@ -1,0 +1,10 @@
+# position_prediction_logic 
+DT = 0.1
+X_THRESHOLD = 0.55
+DUPLICATE_THRESHOLD = 0.06  # Maximaler Abstand in X, um Objekte als Duplikate zu betrachten
+SENTINEL_POSITION = (-1.0, -1.0, -1.0)
+SENTINEL_TYPE = -1
+
+# position_prediction_node
+DEFAULT_CONVEYOR_BELT_SPEED = 0.01
+SENTINEL_MSG = (-1.0, -1.0, -1.0, -1)
